+++ modify_epub/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [1.8.0] - 2022-09-XX
 _All kiwidude plugins updated/migrated to: https://github.com/kiwidude68/calibre_plugins_
+### Added
+- Add keyboard shortcuts access and help to configuration dialog.
 ### Changed
 - **Breaking:** Drop PyQt4 support, require calibre 2.85.1 or later.
 - Refactoring of common code